refactor: extract routers/annotations.py
Move /annotate/* and /report-evaluation/* routes out of main.py into
api/routers/annotations.py. Wire router into main.py.
